Exponenciacao e refatoração

+++ controlador/interpolacao/funcoesInterpolacao.cpp
@@ -40,7 +40,7 @@ void funcaoInterpNewton (std::vector <interp> &pontosInterp, std::vector <interp
 }
 
 void interpNewtonG1Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
@@ -80,7 +80,7 @@ void interpNewtonG1Fechada (std::vector <interp> &pontosInterp, std::vector <int
 }
 
 void interpNewtonG1Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
@@ -127,7 +127,7 @@ void interpNewtonG1Aberta (std::vector <interp> &pontosInterp, std::vector <inte
 }
 
 void interpNewtonG2Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, s2, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, s2, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
@@ -169,7 +169,7 @@ void interpNewtonG2Fechada (std::vector <interp> &pontosInterp, std::vector <int
 }
 
 void interpNewtonG2Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, s2, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, s2, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
@@ -219,7 +219,7 @@ void interpNewtonG2Aberta (std::vector <interp> &pontosInterp, std::vector <inte
 }
 
 void interpNewtonG3Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, s2, s3, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, s2, s3, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
@@ -263,7 +263,7 @@ void interpNewtonG3Fechada (std::vector <interp> &pontosInterp, std::vector <int
 }
 
 void interpNewtonG3Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, s2, s3, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, s2, s3, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
@@ -315,7 +315,7 @@ void interpNewtonG3Aberta (std::vector <interp> &pontosInterp, std::vector <inte
 }
 
 void interpNewtonG4Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, s2, s3, s4, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, s2, s3, s4, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
@@ -361,7 +361,7 @@ void interpNewtonG4Fechada (std::vector <interp> &pontosInterp, std::vector <int
 }
 
 void interpNewtonG4Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
-    float pos, s, s2, s3, s4, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
+    double pos, s, s2, s3, s4, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
     tamParticoes = ((fim - inter[0]) / QTD_DIST_UM);
