+++ controlador/parser/parser.h
@@ -10,7 +10,7 @@
 
 bool analisador (std::string &expressao);
 
-float lerConstante (std::string &expressao, int &percorrer); 
+double lerConstante (std::string &expressao, int &percorrer); 
 std::queue <expre::expre*>* construirFila (std::string &expressao);
 expre::expre* construirArvore (std::queue <expre::expre*> *expressaoPoIn);
 expre::expre* parser (std::string &expressao);
